refactor: 클라이언트 컴포넌트 네이밍 규칙 반영 (#10)
