+++ apps/web/src/app/mypage/_components/BannerContainer/BannerContainer.tsx
@@ -1,17 +1,17 @@
 import { Suspense } from 'react';
 import BannerFallback from './_components/BannerFallback/BannerFallback';
 import styles from './BannerContainer.module.css';
-import DdayBannerContainer from '@/app/mypage/_components/BannerContainer/_clientBoundary/DdayBannerContainer/DdayBannerContainer';
-import TotalPhotoCountBannerContainer from '@/app/mypage/_components/BannerContainer/_clientBoundary/TotalPhotoCountBannerContainer/TotalPhotoCountBannerContainer';
+import DdayBannerContainerClient from '@/app/mypage/_components/BannerContainer/_clientBoundary/DdayBannerContainerClient/DdayBannerContainerClient';
+import TotalPhotoCountBannerContainerClient from '@/app/mypage/_components/BannerContainer/_clientBoundary/TotalPhotoCountBannerContainerClient/TotalPhotoCountBannerContainerClient';
 
 export default function BannerContainer() {
   return (
     <section className={styles.wrapper}>
       <Suspense fallback={<BannerFallback />}>
-        <DdayBannerContainer />
+        <DdayBannerContainerClient />
       </Suspense>
       <Suspense fallback={<BannerFallback />}>
-        <TotalPhotoCountBannerContainer />
+        <TotalPhotoCountBannerContainerClient />
       </Suspense>
     </section>
   );
