+++ apps/web/src/app/mypage/_components/BannerContainer/_clientBoundary/TotalPhotoCountBannerContainerClient/TotalPhotoCountBannerClient.tsx
